fix: latest tweets retrieval optimization is unpredictable
`getTweets` does not always return results in a predictable order, which
means that the first tweet in not necessarily the latest one. This
breaks the synchronization optimization because it relies on this
assumption.

This fix reverses the implementation for this optimization: instead of
preventing the sync when the latest tweet if cached, we prevent the
sync by default but allow it to run if we find that *any* of the
${LATEST_TWEETS_COUNT} tweets is not cached.